arm64: limpa os quatro avisos do recompilador, sem mudar comportamento (AI-assisted)
JUSTIFICATIVA DE REGRA 1 (FORK.md): isto toca pcsx2/arm64/, caminho
protegido. Entra sem A/B porque nao ha comportamento para medir — uma
mudanca e anotacao pura e a outra foi provada inerte por enumeracao. O
gate phase-0.5 de correcao ARM64 continua sendo o juiz.

Os quatro avisos vieram da build da alpha 21, de graca:

1-2) VuFmacFlags-arm64.h:190 e :309, -Wunused-variable. O corpo do laco e
     so um pxAssert, que desaparece em release e deixa a variavel do
     range-for sem uso. Codigo correto, aviso legitimo. Resolvido com
     [[maybe_unused]] na propria variavel, que preserva a asserçao em
     debug em vez de apagar a verificacao para calar o compilador.

3-4) microVU_Compile-arm64.inl:451 e :466, -Wsign-compare, no detector de
     spin loop da VU0. Parecia perigoso e nao era: `startPC` e u32, entao
     `startPC + 8 + imm11 * 8` ja e u32 por promocao, e comparar u32 com
     `(s32)startPC` converteria o s32 de volta para u32. O cast nunca fez
     nada alem de gerar o aviso.

     Nao confiei no raciocinio: enumerei todos os pares (startPC de 0 a
     16 KiB em passo de 8, imm11 de -1024 a 1023) comparando com e sem o
     cast. ZERO divergencias. E o caso de subfluxo continua correto —
     pc=0 com imm=-2 da alvo 0xfffffff8, que nao casa, que e a resposta
     certa para "isto nao e um auto-laco".

Nenhum destes e otimizacao. Sao ruido que estava escondendo avisos futuros
de verdade: quatro linhas constantes num log de build sao quatro linhas que
ninguem le mais, e o proximo aviso real chega no meio delas.

Registro de processo, porque me pegou agora: fork-diff.sh compara
BASE..HEAD, entao rodar antes de commitar descreve o commit ANTERIOR, nao a
mudanca pendente. Ele me disse "OK" com estas edicoes na arvore. O uso
correto e depois do commit.

+++ pcsx2/arm64/microVU_Compile-arm64.inl
@@ -448,7 +448,11 @@ static mVUSpinLoop mVUdetectSpinLoop(mV, u32 startPC)
 	r.viA = (lo0 >> 16) & 0x1F;
 	r.viB = (lo0 >> 11) & 0x1F;
 	const s32 imm11 = (s32)(lo0 << 21) >> 21;
-	if (startPC + 8 + imm11 * 8 == (s32)startPC)
+	// Sem o (s32) que havia aqui, e o cast era inerte: `startPC` e u32, entao o lado esquerdo
+	// ja e u32 por promocao, e comparar u32 com s32 converteria o s32 de volta para u32. So
+	// gerava -Wsign-compare. A aritmetica com envolvimento e a pretendida: um alvo que
+	// subflui (pc=0, imm=-2 -> 0xfffffff8) simplesmente nao casa, que e a resposta certa.
+	if (startPC + 8 + imm11 * 8 == startPC)
 	{
 		// 2-pair self-loop: spins while taken; exit on the opposite sense.
 		r.found = true;
@@ -463,7 +467,7 @@ static mVUSpinLoop mVUdetectSpinLoop(mV, u32 startPC)
 	if ((lo2 >> 25) != 0x20) // B
 		return r;
 	const s32 imm2 = (s32)(lo2 << 21) >> 21;
-	if (startPC + 16 + 8 + imm2 * 8 != (s32)startPC)
+	if (startPC + 16 + 8 + imm2 * 8 != startPC)
 		return r;
 	r.found = true;
 	r.exitOnEq = ibeq; // spins on fallthrough; taken (the branch sense) exits
